Accept `size_t` in `lauf_asm_declare_block()`
This matches the return type of `lauf_asm_build_get_vstack_size()`.

+++ include/lauf/asm/builder.h
@@ -74,7 +74,7 @@ typedef struct lauf_asm_block lauf_asm_block;
 ///
 /// It is only valid inside the current function.
 /// If it's the first block, it becomes the entry block.
-lauf_asm_block* lauf_asm_declare_block(lauf_asm_builder* b, uint8_t input_count);
+lauf_asm_block* lauf_asm_declare_block(lauf_asm_builder* b, size_t input_count);
 
 /// Sets the insertion point of the builder to append instruction to the end of the block.
 ///
